Movement and Map Reader
The movement has been setup for the camera to rotate around the object. I want to do it was momentum since collision detection using translation is rather bad, but I don't know if I'll have time for that.
Furthermore, I now have a map reader to bring in the maze. This should make it easy to make mazes. The idea is that it will be broken down into at least 5 mazes, which maze start acting as a checkpoint. When the player finishes one maze, they switch over to another one. I probably won't have time for bonus stuff, but I might be able to finish up the main game.